[FEATURE] Adding native support for CC and Reviewer
Adding reviewers and CC users in a Gerrit server
environment shall no longer happen via the
git receive-pack approach (deprecated approach),
but instead should be used via an extension to
the rhs of the refspec.
This commit introduced additional script parameters,
which allow to specify multiple CC users and/or
multiple Reviewers, which will be passed via the
rhs of the refspec.
Corresponding tests are added to the unit test
suite.